[core] Correctly test column visibility switch impact on column width
