+++ packages/grid/data-grid/src/tests/layout.DataGrid.test.tsx
@@ -378,61 +378,6 @@ describe('<DataGrid /> - Layout & Warnings', () => {
         });
       });
 
-      it('should resize flex: 1 column when setting hide: false on a column to avoid exceeding grid width', async () => {
-        const TestCase: React.FC<DataGridProps> = (props) => (
-          <div style={{ width: 300, height: 500 }}>
-            <DataGrid {...props} />
-          </div>
-        );
-
-        const { setProps } = render(
-          <TestCase
-            rows={[
-              {
-                id: 1,
-                first: 'Mike',
-                age: 11,
-              },
-              {
-                id: 2,
-                first: 'Jack',
-                age: 11,
-              },
-              {
-                id: 3,
-                first: 'Mike',
-                age: 20,
-              },
-            ]}
-            columns={[
-              { field: 'id', flex: 1 },
-              { field: 'first', width: 100 },
-              { field: 'age', width: 50, hide: true },
-            ]}
-          />,
-        );
-
-        let firstColumn = document.querySelector('[role="columnheader"][aria-colindex="1"]');
-        // @ts-expect-error need to migrate helpers to TypeScript
-        expect(firstColumn).toHaveInlineStyle({
-          width: '198px', // because of the 2px border
-        });
-
-        setProps({
-          columns: [
-            { field: 'clientId', flex: 1 },
-            { field: 'first', width: 100 },
-            { field: 'age', width: 50 },
-          ],
-        });
-
-        firstColumn = document.querySelector('[role="columnheader"][aria-colindex="1"]');
-        // @ts-expect-error need to migrate helpers to TypeScript
-        expect(firstColumn).toHaveInlineStyle({
-          width: '148px', // because of the 2px border
-        });
-      });
-
       it('should be rerender invariant', () => {
         const Test = () => {
           const columns = [{ field: 'id', headerName: 'ID', flex: 1 }];
