+++ packages/grid/x-grid/src/tests/layout.XGrid.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { createClientRenderStrictMode } from 'test/utils';
 import { expect } from 'chai';
-import { XGrid } from '@material-ui/x-grid';
+import {GridApiRef, useGridApiRef, XGrid} from '@material-ui/x-grid';
 
 describe('<XGrid /> - Layout', () => {
   // TODO v5: replace with createClientRender
@@ -80,4 +80,60 @@ describe('<XGrid /> - Layout', () => {
       });
     });
   });
+
+  describe('columns width', () => {
+    it('should resize flex: 1 column when changing column visibility to avoid exceeding grid width', () => {
+      let apiRef: GridApiRef
+
+      const TestCase = (props) => {
+        apiRef = useGridApiRef();
+
+        return (
+            <div style={{ width: 300, height: 500 }}>
+              <XGrid {...props} apiRef={apiRef} />
+            </div>
+        );
+      };
+
+      render(
+          <TestCase
+              rows={[
+                {
+                  id: 1,
+                  first: 'Mike',
+                  age: 11,
+                },
+                {
+                  id: 2,
+                  first: 'Jack',
+                  age: 11,
+                },
+                {
+                  id: 3,
+                  first: 'Mike',
+                  age: 20,
+                },
+              ]}
+              columns={[
+                { field: 'id', flex: 1 },
+                { field: 'first', width: 100 },
+                { field: 'age', width: 50, hide: true },
+              ]}
+          />,
+      );
+
+      let firstColumn = document.querySelector('[role="columnheader"][aria-colindex="1"]');
+      // @ts-expect-error need to migrate helpers to TypeScript
+      expect(firstColumn).toHaveInlineStyle({
+        width: '198px', // because of the 2px border
+      });
+
+      apiRef!.current.setColumnVisibility('age', true);
+      firstColumn = document.querySelector('[role="columnheader"][aria-colindex="1"]');
+      // @ts-expect-error need to migrate helpers to TypeScript
+      expect(firstColumn).toHaveInlineStyle({
+        width: '148px', // because of the 2px border
+      });
+    });
+  })
 });
